requirements.txt
Django
djangorestframework
Pillow
gunicorn
whitenoise
dj-database-url
psycopg2-binary
